Add description list & fix list-in-paragraph error
Added description list described in Additional Resources link:
https://learn.shayhowe.com/html-css/creating-lists/#description-lists
Copied sample list definition, and interspersed with comments describing
proper and possible usage.

Ran file through validator.w3.org/nu, and learned that lists should not
(in fact, cannot) be nested inside paragraphs. If they are, the browser
will treat the beginning of the list as the end of the paragraph, and
ignore the closing paragraph tag, and render them accordingly.

Corrected the nested list in paragraph error, and added comment block
describing the rule, with URLs to relevant documentation.